requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
SQLAlchemy==2.0.29
pydantic==2.6.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
